+++ src/app/controllers/index.ts
@@ -1,6 +1,5 @@
 import "./export.controller";
 import "./login.controller";
-import "./emuwebapp.controller";
 import "./manual.controller";
 import "./modal.controller";
 import "./search-annot.controller";